refactor(routing): extract shared validation to common module
Extract duplicate routing validation logic (~60 lines) from sync and
async RouteNode executors into a shared module. Both executors now
import from routing_validation.py.
